refactor(ui): split RuntimeControlPage subcomponents into pages/runtime/
RuntimeControlPage.tsx was 589 lines; over half was three subcomponents
that don't share state with the page's Start/Stop flow. Pull them out
so the page reads as a clean lifecycle controller and each piece is
maintainable in isolation.

Layout:
- RuntimeControlPage.tsx (363)              — page shell + start/stop flow
- pages/runtime/RunningSimulationCard.tsx (129) — green 'live' card with
                                              StatBlocks + download YAML
- pages/runtime/GlobalDebugLevelCard.tsx (94)   — apply one DebugLevel
                                              to every protocol
- pages/runtime/AdvancedSection.tsx (28)         — collapsed <details>
                                              wrapping the debug card

Pure code move; behaviour, props, and DOM shape are unchanged. biome /
tsc / vitest (65 tests) all pass; full make build produces a working
bundle.
